+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 `commitmsgfmt` formats commit messages. It reflows and wraps text, with special
 understanding of patterns often seen in commit messages.
 
+## Unreleased
+
+- #1: Fix continuation indentation of tab-indented list items
+
 ## 1.0.0 - 2018-05-09
 
 - First public release of `commitmsgfmt`; the "make it work" version.